pyproject.toml
[project]
name = "cognito-jwt-verifier"
version = "0.0.1"
description = "A utility for verifying JWTs issued by AWS Cognito."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest==8.*",
    "aiohttp",
    "pyjwt[crypto]",
]
